Move add-session spec to availability.spec.ts FT tests
